+++ akash/client/lease.go
@@ -9,7 +9,7 @@ func (ak *AkashClient) CreateLease(seqs Seqs, provider string) (string, error) {
 		SetDseq(seqs.Dseq).SetGseq(seqs.Gseq).SetOseq(seqs.Oseq).
 		SetProvider(provider).SetOwner(ak.Config.AccountAddress).SetFrom(ak.Config.KeyName).
 		DefaultGas().SetChainId(ak.Config.ChainId).SetKeyringBackend(ak.Config.KeyringBackend).
-		AutoAccept().SetNode(ak.Config.Node).OutputJson()
+		SetNote(ak.transactionNote).AutoAccept().SetNode(ak.Config.Node).OutputJson()
 
 	out, err := cmd.Raw()
 	if err != nil {