+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Transaction memo stating the Terraform provider was used and which version
 ### Fixed
 - Bug where destroying deployments did not really close them
+- Owner not being set in Go client for bid listing
 ### Development
 - Replaced HTTP API call for deployments with CLI